Separate the storage of lyric "source path" and "save path"
This is necessary for correctly deleting the currently-loaded lyrics.
It also clears up some ambiguity of what the "persistent storage path"
is and when it is set. Source path is always set (but might not point to
something locally accessible, it could be a URL), save path is only set
if the lyrics have been saved.