Switch from libxml2 to pugixml for html parsing in web page sources
This is mostly because libxml2 is GPL and therefore (as far as I know)
is incompatible with our own MIT license, so I'm swapping it out.

pugixml alone is not capable of properly parsing html (see https://github.com/zeux/pugixml/issues/106),
but we have libtidy which converts the raw html page into xhtml which it
apparently *is* capable of properly parsing. So we should be ok.

IMPORTANT NOTE: This commit will not compile because the code changes to
remove usage of libxml are not included. They will be included instead
in the next commit, just for visibility (so that those changes aren't
lost in this several-thousand-file "change").