Add (experimental) support for showing the panel in an external window
This is experimental for this release, and will be solidified in the
following release. It's experimental just because it's a large change
that involves completely rewriting the rendering logic. As a result
the rendering code has been duplicated so that the external window uses
Direct2D while the normal panel continues to use GDI.

The next release will re-unify the two codepaths so that everything uses
Direct2D, but I'm keeping this one "experimental" so as to not break
everybody's panel immediately if the Direct2D implementation is broken
in some significant way.

Resolves #41